requirements.txt
Flask==2.3.0
SQLAlchemy==2.0.15
PyMySQL==1.1.0
python-dotenv==1.0.0